Move Module Manager from decodex.data to decodex.modules
